Handle overflow of "commitment"
(cherry picked from commit e7210a6a38e46dfc7acdddf9ca20da7183ba8fcd)

+++ app/src/main/java/com/android/wcf/model/LeaderboardTeam.kt
@@ -12,7 +12,7 @@ data class LeaderboardTeam(
         @SerializedName("creator_id") var leaderId: String = "",
         @SerializedName("hidden") var hidden: Boolean = false,
         @SerializedName("distance") var distance: Int = 0,
-        @SerializedName("commitment") var commitment: Int = 0
+        @SerializedName("commitment") var commitment: Long = 0
 ) : Parcelable {
     var rank: Int = 0
     var amountPledged: Double = 0.0
